LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/structured_clone_bigint-expected.txt

CONSOLE MESSAGE: SyntaxError: No identifiers allowed directly after numeric literal

Harness Error (FAIL), message = SyntaxError: No identifiers allowed directly after numeric literal
